Project#2
Moving the Users Code right to the User.js file
